First version of evaluation script
In its main mode, the script tries out all combinations of the --no...
options (there are currently five of these options, so 32 combinations)
and measures the running time for each, on a given input.

In analyze mode, the result is analyzed and the effect of each option
(relative to the others) is printed on the screen.